requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
azure-devops
azure-identity>=1.13.0
google-generativeai>=0.3.1
google-genai
Flask
Flask-RESTful
flask_cors
openai
redis